+++ pkg/rancher-components/src/components/RcDropdown/index.ts
@@ -1,0 +1,5 @@
+export { default as RcDropdown } from './RcDropdown.vue';
+export { default as RcDropdownCollection } from './RcDropdownCollection.vue';
+export { default as RcDropdownItem } from './RcDropdownItem.vue';
+export { default as RcDropdownSeparator } from './RcDropdownSeparator.vue';
+export { default as RcDropdownTrigger } from './RcDropdownTrigger.vue';
