+++ pkg/rancher-components/src/index.ts
@@ -2,4 +2,11 @@ export { Card } from './components/Card';
 export { BadgeState } from './components/BadgeState';
 export { Banner } from './components/Banner';
 export { StringList } from './components/StringList';
+export {
+  RcDropdown,
+  RcDropdownCollection,
+  RcDropdownItem,
+  RcDropdownSeparator,
+  RcDropdownTrigger
+} from './components/RcDropdown';
 export * from './components/Form';
